Fix clippy warnings

+++ src/lib.rs
@@ -119,7 +119,7 @@ mod tests {
         dbg!(&chunk_footers);
         assert_eq!(
             chunk_footers.len(),
-            ((record_count + RECORD_COUNT - 1) / RECORD_COUNT) as usize,
+            (record_count + RECORD_COUNT - 1) / RECORD_COUNT,
             "number of chunk footers must be equal to the number of chunks, which is the number of records divided by the record count"
         );
         for (i, chunk_footer) in chunk_footers.iter().enumerate() {
